+++ pkg/vm/engine/disttae/txn_table.go
@@ -560,7 +560,7 @@ func (tbl *txnTable) resetSnapshot() {
 func (tbl *txnTable) CollectTombstones(
 	ctx context.Context, txnOffset int,
 ) (engine.Tombstoner, error) {
-	tombstone := buildTombstoneWithDeltaLoc()
+	tombstone := NewEmptyTombstoneWithDeltaLoc()
 
 	offset := txnOffset
 	if tbl.db.op.IsSnapOp() {
